Refactor Dockerfile, update riak.conf
Refactor the Dockerfile to take better advantage of cache layers,
particularly for package installs and around the Erlang install as much
as possible. This also copies the Riak LICENSE file into the
/usr/local/riak directory.

Update the riak.conf to detail a couple other changed fields. There's
no value changes in this, just remembering there were a few additional
values changed and it'd be good to cover those in the comment.